requirements_cloud_run.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
librosa==0.10.1
tensorflow==2.13.0
scikit-learn==1.3.0
google-cloud-storage==2.10.0
pydantic==2.5.0